Survival-Analytics---Kaplan-Meier-Estimations
Developed Kaplan Meier Estimation and LogRank test algorithms from scratch. 
Kaplan Meier Estimation is a non parametric survival analysis algorithm used to estimate the
survival probabailities. LogRank test is a statistical test that is used for comparing the 
survival distributions of two or more groups
In this project, I estimated and compared the survival probabilities of male and female groups of lung data set which
is present in the 'Survival' package